Make actors actor+session sepecific
This commit makes the actors variable dependent on each actors, as
opposed to being session specific. This is the initial step to perfoming
man-in-the-middle attacks!

This is done by introducing the concept of kinds to variables and
constants, such that variables only can be unified to messages of the
same kind. Doing so removes a lot of duplication, and simplifies the
logic as well.

In addition to this, a `ForWho` parameter is now passed to the converter
making it more clear who each variable/actor/constant is registered for.