Attempt to break cycles due to unification in can_derive
This occured when verifing TLS_1_3 for one session. Somehow a
unification resulted in a cycle in a term, thus making can_derive run
without stopping. The currect mitigation is to stop if the stack grows
too large, and return that the term can't be unified, hopefully ending
that branch of execution.